Extract InvitationService from MessageDB (5 functions, ~1,204 lines)
