+++ .readme/tasks/messagedb-refactoring-analysis.md
@@ -25,11 +25,6 @@ The `src/components/context/MessageDB.tsx` file has grown to 5,650 lines and vio
 
 **📋 Comprehensive Test Documentation**: See [`src/dev/refactoring/test-implementation-guide.md`](../src/dev/refactoring/test-implementation-guide.md) for complete testing strategy, test status, and incremental workflow procedures.
 
-**🎯 Current Test Status**:
-- ✅ **61 tests passing** across 5 test files (3.05s execution)
-- ✅ **7 critical functions** protected with mock integration tests
-- ✅ **Immediate failure detection** for breaking changes during refactoring
-- ✅ **Phase 1 Complete** - Ready for service extraction
 
 ## 📊 EXTRACTION PROGRESS TRACKER
 
@@ -88,21 +83,16 @@ const functionName = React.useCallback(async (...) => {
 - [x] **informSyncData** - 63 lines (line 828-890)
 
 #### ConfigService (to be created - 2 functions, ~355 lines)
-- [ ] **getConfig** - 261 lines (line 2907-3167)
-- [ ] **saveConfig** - 94 lines (line 3169-3262)
+- [x] **getConfig** - 261 lines (line 2907-3167)
+- [x] **saveConfig** - 94 lines (line 3169-3262)
 
 #### InvitationService (to be created - 5 functions, ~1,204 lines)
-- [ ] **constructInviteLink** - 60 lines (line 1973-2032)
-- [ ] **sendInviteToUser** - 28 lines (line 2005-2032)
-- [ ] **generateNewInviteLink** - 374 lines (line 2034-2407)
-- [ ] **processInviteLink** - 399 lines (line 2409-2807)
-- [ ] **joinInviteLink** - 343 lines (line 2465-2807)
+- [x] **constructInviteLink** - 60 lines (line 1973-2032)
+- [x] **sendInviteToUser** - 28 lines (line 2005-2032)
+- [x] **generateNewInviteLink** - 374 lines (line 2034-2407)
+- [x] **processInviteLink** - 399 lines (line 2409-2807)
+- [x] **joinInviteLink** - 343 lines (line 2465-2807)
 
-
-**📊 Extraction Progress Summary:**
-- ✅ **Completed**: 2 services, 8 functions, ~2,577 lines extracted
-- ⏳ **Remaining**: 5 services, 26 functions, ~3,427 lines to extract
-- 🎯 **Next**: SpaceService (7 functions, ~1,130 lines)
 
 **🚨 Testing Requirements During Refactoring**:
 1. **NEVER skip testing** after each service extraction
@@ -332,6 +322,8 @@ Before extracting anything, perform complete audit:
 **✅ Integration Complete Criteria**: MessageDB.tsx now delegates to service, API unchanged
 
 ### **Step 5: TEST AND VERIFY**
+
+Step 5 can be run when ALL functions have been extracted, IF those functions where small.
 
 1. **Run all tests**: `yarn test` - ALL tests must pass
 2. **Check TypeScript**: `yarn tsc --noEmit` - no compilation errors
